test(auth): split test_auth.py into 6 cluster files (D1 PR-2 baseline)
Structural split only — test bodies preserved verbatim from the original
test_auth.py monolith. Subsequent commits in this PR migrate the bodies
from monkeypatch.setattr / core.X = AsyncMock patterns to constructor
injection via tests/_fixtures/make_fake_core(), delete the _AuthFacadeModule
write-through facade from auth.py, and update ADR-003 + lint allowlist.

Cluster mapping per phase-1.md#arch-d1-auth-side:
  storage:        508 LOC — LoadAuthFromStorage, LoadHttpxCookies*, AuthTokensFromStorage, LoaderFlatCookieParity, etc.
  extraction:     648 LOC — ExtractCookies, ExtractCSRF, ExtractSessionId, FinalUrlScrubbing, edge cases
  refresh:        721 LOC — FetchTokens, FetchTokensAutoRefresh
  cookie_policy: 1433 LOC — IsAllowedCookieDomain*, RookiepyDomainsCoverage, regional/path-aware identity
  keepalive:      642 LOC — IsRecentlyRotated, PokeConcurrencyThrottling, KeepalivePoke
  account:        397 LOC — ExtractEmailFromHtml, EnumerateAccounts, AccountMetadata, AuthuserPlumbing

Net delta: +259 LOC across 6 files from duplicated imports / pytest plumbing.
Test count: 297 collected (original had 149 def test_*; expansion from parametrize is preserved).
